+++ src/main/java/com/deepoove/poi/plugin/toc/TOCRenderPolicy.java
@@ -13,24 +13,35 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.deepoove.poi.policy;
+package com.deepoove.poi.plugin.toc;
 
+import org.apache.poi.xwpf.usermodel.XWPFParagraph;
 import org.apache.poi.xwpf.usermodel.XWPFRun;
+import org.openxmlformats.schemas.wordprocessingml.x2006.main.CTP;
+import org.openxmlformats.schemas.wordprocessingml.x2006.main.CTSimpleField;
+import org.openxmlformats.schemas.wordprocessingml.x2006.main.STOnOff;
 
 import com.deepoove.poi.XWPFTemplate;
+import com.deepoove.poi.policy.RenderPolicy;
 import com.deepoove.poi.template.ElementTemplate;
 import com.deepoove.poi.template.run.RunTemplate;
-import com.deepoove.poi.xwpf.BodyContainer;
-import com.deepoove.poi.xwpf.BodyContainerFactory;
 
-public class CommentPolicy implements RenderPolicy {
+/**
+ * Experimental: Table of contents
+ */
+public class TOCRenderPolicy implements RenderPolicy {
 
     @Override
     public void render(ElementTemplate eleTemplate, Object data, XWPFTemplate template) {
-        RunTemplate runTemplate = (RunTemplate) eleTemplate;
-        XWPFRun run = runTemplate.getRun();
-        BodyContainer bodyContainer = BodyContainerFactory.getBodyContainer(run);
-        bodyContainer.clearPlaceholder(run);
+        XWPFRun run = ((RunTemplate) eleTemplate).getRun();
+        run.setText("", 0);
+
+        XWPFParagraph tocPara = (XWPFParagraph) run.getParent();
+        CTP ctP = tocPara.getCTP();
+
+        CTSimpleField toc = ctP.addNewFldSimple();
+        toc.setInstr("TOC \\o");
+        toc.setDirty(STOnOff.TRUE);
     }
 
 }
