--- src/main/java/com/deepoove/poi/policy/TOCRenderPolicy.java
@@ -1,46 +1,0 @@
-/*
- * Copyright 2014-2020 Sayi
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.deepoove.poi.policy;
-
-import org.apache.poi.xwpf.usermodel.XWPFParagraph;
-import org.apache.poi.xwpf.usermodel.XWPFRun;
-import org.openxmlformats.schemas.wordprocessingml.x2006.main.CTP;
-import org.openxmlformats.schemas.wordprocessingml.x2006.main.CTSimpleField;
-import org.openxmlformats.schemas.wordprocessingml.x2006.main.STOnOff;
-
-import com.deepoove.poi.XWPFTemplate;
-import com.deepoove.poi.template.ElementTemplate;
-import com.deepoove.poi.template.run.RunTemplate;
-
-/**
- * Experimental: Table of contents
- */
-public class TOCRenderPolicy implements RenderPolicy {
-
-    @Override
-    public void render(ElementTemplate eleTemplate, Object data, XWPFTemplate template) {
-        XWPFRun run = ((RunTemplate) eleTemplate).getRun();
-        run.setText("", 0);
-
-        XWPFParagraph tocPara = (XWPFParagraph) run.getParent();
-        CTP ctP = tocPara.getCTP();
-
-        CTSimpleField toc = ctP.addNewFldSimple();
-        toc.setInstr("TOC \\o");
-        toc.setDirty(STOnOff.TRUE);
-    }
-
-}
